requirements.txt
# Auto-Coin Dependencies
yfinance>=0.2.28
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
streamlit>=1.28.0
plotly>=5.15.0
pyyaml>=6.0
pyarrow>=14.0.0

